requirements.txt
django
djangorestframework
nltk==3.7
pandas==1.4.2
xlsxwriter
mysqlclient
django-cors-headers
pytesseract # along with Tesseract to be installed separately
geotext
nltk